Automatically populate all settings values with defaults
This makes it so users only need to copy the `plugin.tcl` file onto the
tablet.  The plugin will now automatically populate the settings with
default values, even if `settings.tdb` is not present.  The old behavior
caused confusing errors for users that only copied on `plugin.tcl`
without `settings.tdb`