Retract HA auto-discovery messages when feature is disabled
If the HA auto-discovery feature was enabled and is then turned off,
attempt to automatically retract all retained messages we had published
to the MQTT broker about our settings.  This makes it easier to clean up
this retained state on the MQTT broker without having manually clear it
out.

This functionality isn't perfect: if you have changed the auto-discovery
prefix or the device unique ID then the retractions will be sent with
the new values rather than the old ones.  However, this should still be
useful in the majority of cases.  These issues could be fixed if we
kept separate copies of the old and new settings values, but that
requires slightly more refactoring that doesn't seem worthwhile right
now.