Support reporting and controlling the steam heater state
This adds 2 new fields to the reported state:
- `steam_mode` which reports the steam mode as "On", "Off", or "Eco"
- `steam_state` which reports the steam state as a true/false boolean,
  for easy use in a Home Assistant switch.  The "Eco" state is treated
  as on.

This also adds 2 new commands:
- `steam_on` this turns the steam heater on, and resets the eco mode
  timer if steam eco mode is enabled.
- `steam_off` this turns the steam heater off.

When Home Assistant auto-discovery is enabled, a new switch is published
to HA to allow controlling the steam heater state.  If Eco steam mode is
enabled under the machine calibration settings, an additional
"Steam Mode" entity will be published as well, which reports the state
as On/Off/Eco.